Move logging to separate class to prevent allocation per generic type
